reorganize project structure
add JsonUnit to project
