+++ .gitignore
@@ -1,5 +1,5 @@
 # Ignore Gradle project-specific cache directory
-.gradle
+core/.gradle
 
 # Ignore Gradle build output directory
 build
