+++ build.gradle
@@ -1,11 +1,3 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Groovy project to get you started.
- * For more details take a look at the Groovy Quickstart chapter in the Gradle
- * User Manual available at https://docs.gradle.org/5.6.2/userguide/tutorial_groovy_projects.html
- */
-
 plugins {
     // Apply the groovy plugin to add support for Groovy
     id 'groovy'
@@ -22,7 +14,10 @@ dependencies {
     implementation 'org.codehaus.groovy:groovy-all:2.5.7'
     // Use the awesome Spock testing and specification framework
     implementation 'org.spockframework:spock-core:1.3-groovy-2.5'
-
-    // Use the awesome Spock testing and specification framework
-    testImplementation 'org.spockframework:spock-core:1.3-groovy-2.5'
+    // https://mvnrepository.com/artifact/org.hamcrest/hamcrest
+    implementation 'org.hamcrest:hamcrest:2.2'
+    // https://mvnrepository.com/artifact/org.springframework/org.springframework.test
+    implementation 'org.springframework:org.springframework.test:3.2.2.RELEASE'
+    // json assert lib
+    implementation 'net.javacrumbs.json-unit:json-unit:2.21.0'
 }
