--- core/.gitignore
@@ -1,6 +1,0 @@
-# Ignore Gradle project-specific cache directory
-.gradle
-
-# Ignore Gradle build output directory
-build
-/.idea/
